Rename misspelled file.
