+++ src/features/account/pages/VerifyPage.tsx
@@ -7,8 +7,8 @@ import ZUILogoLoadingIndicator from 'zui/ZUILogoLoadingIndicator';
 import ZUIText from 'zui/components/ZUIText';
 import ZUIButton from 'zui/components/ZUIButton';
 import { Msg, useMessages } from 'core/i18n';
-import messageIds from '../l10n/messagesIds';
-import messagesIds from '../l10n/messagesIds';
+import messageIds from '../l10n/messageIds';
+import messagesIds from '../l10n/messageIds';
 import { useSendVerification } from '../hooks/useSendVerification';
 import ZUIAlert from 'zui/components/ZUIAlert';
 import { SendFail, SendVerificationStatus } from '../types';
